requirements.txt
apache-airflow==2.10.3
pandas==2.2.3
requests==2.31.0
